discord-bot/requirements.txt
discord.py-self>=2.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
